+++ common/__init__.py
@@ -1,6 +1,6 @@
 # API Python wrapper for The Next Generation Vulnerability & Threat Intelligence Database  - https://vfeed.io
-# Copyright (C) 2013 - 2018 vFeed IO
+# Copyright (C) 2013 - 2019 vFeed IO
 
-__title__ = "Python 3.x API For vFeed Professional Editions"
-__version__ = "Beta 0.9.7"
+__title__ = "Python 3.x API For vFeed Professional Edition"
+__version__ = "0.9.8"
 __email__ = "support@vfeed.io"